Revert "Waypoint navigation through built-in mission (#64)"
This reverts commit 6a1116d3ab5dd8a62124ae0d5e3d07df7333e19c.
